// backend/src/Model/user.ts
type User = {
    id: number;
    firstName: string;
    lastName: string;
    email: string;
    password: string;
    phone: string;
    isAdmin: number;
    isHeadCoach: number;
    isAssistantCoach: number;
    isCoach: number;
    isMentor: number;
    isStudent: number;
    isActive: number;
    leetcodeUsername: string;
    codeforcesUsername: string;
    uhuntId: number;
  }

  export { User }